+++ lib/src/data_frame/factories/from_matrix.dart
@@ -1,6 +1,6 @@
 import 'package:ml_dataframe/ml_dataframe.dart';
-import 'package:ml_dataframe/src/data_frame/data_frame_helpers.dart';
 import 'package:ml_dataframe/src/data_frame/data_frame_impl.dart';
+import 'package:ml_dataframe/src/data_frame/helpers/get_header.dart';
 import 'package:ml_dataframe/src/numerical_converter/numerical_converter_impl.dart';
 import 'package:ml_linalg/linalg.dart';
 import 'package:quiver/iterables.dart';
@@ -21,13 +21,10 @@ DataFrame fromMatrix(Matrix data, {
       ? data.sample(columnIndices: columns)
       : data;
 
-  final areSeriesDiscrete = enumerate(header).map((indexedName) {
-    if (discreteColumns.contains(indexedName.index) ||
-        discreteColumnNames.contains(indexedName.value)) {
-      return true;
-    }
-    return false;
-  });
+  final areSeriesDiscrete = enumerate(header).map((indexedName) =>
+    discreteColumns.contains(indexedName.index) ||
+        discreteColumnNames.contains(indexedName.value),
+  );
 
   return DataFrameImpl.fromMatrix(selectedData, header,
       NumericalConverterImpl(false), areSeriesDiscrete);