+++ lib/src/data_frame/helpers/convert_series_to_rows.dart
@@ -1,0 +1,13 @@
+import 'package:ml_dataframe/src/data_frame/series.dart';
+
+Iterable<Iterable<dynamic>> convertSeriesToRows(Iterable<Series> series) sync* {
+  final iterators = series
+      .map((series) => series.data.iterator)
+      .toList(growable: false);
+
+  while (iterators.fold(true, (isActive, iterator) => iterator.moveNext())) {
+    yield iterators
+        .map((iterator) => iterator.current)
+        .toList(growable: false);
+  }
+}