+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.4.1
-- `DataFrame`: 
+- `DataFrame`: series-to-rows and rows-to-series converters issues fixed 
 
 ## 0.4.0
 - `DataFrame`: `fromRawCsv` constructor added